factored out common contract test code
