+++ contracts/test/ligo.ts
@@ -3,6 +3,10 @@ import * as fs from 'fs';
 import * as path from 'path';
 import { $log } from '@tsed/logger';
 
+import { ContractAbstraction } from '@taquito/taquito/dist/types/contract';
+import { ContractProvider } from '@taquito/taquito/dist/types/contract/interface';
+import { TezosToolkit, MichelsonMap } from '@taquito/taquito';
+
 export class LigoEnv {
   readonly cwd: string;
   readonly srcDir: string;
@@ -31,6 +35,10 @@ function defaultLigoEnv(): LigoEnv {
   const out = path.join(cwd, 'out');
   return new LigoEnv(cwd, src, out);
 }
+
+export type Contract = ContractAbstraction<ContractProvider>;
+
+export type address = string;
 
 export async function compileAndLoadContract(
   env: LigoEnv,
@@ -65,4 +73,30 @@ async function runCmd(cwd: string, cmd: string): Promise<void> {
       errout ? reject(errout) : resolve()
     )
   );
+}
+
+export async function originateContract(
+  tz: TezosToolkit,
+  code: string,
+  storage: any,
+  name: string
+): Promise<Contract> {
+  try {
+    const originationOp = await tz.contract.originate({
+      code,
+      init: storage
+    });
+
+    //A HACK
+    // await delay(5000);
+
+    const contract = await originationOp.contract();
+    $log.info(`originated contract ${name} with address ${contract.address}`);
+    $log.info(`consumed gas: ${originationOp.consumedGas}`);
+    return Promise.resolve(contract);
+  } catch (error) {
+    const jsonError = JSON.stringify(error, null, 2);
+    $log.fatal(`${name} origination error ${jsonError}`);
+    return Promise.reject(error);
+  }
 }
